# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Laravel",
        "description": "This is a sample Laravel server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Localhost API Mocking"
        }
    ],
    "paths": {
...
        "/api/addresses/districts": {
            "get": {
                "tags": [
                    "AddressDistrict"
                ],
                "summary": "Get all address district",
                "description": "This request is used to get all address district when user use the App. This request is using MySql database.",
                "operationId": "25ca2820bec0731171b738ea844095c6",
                "responses": {
                    "200": {
                        "description": "Address district fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Get List Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AddressDistrict"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Protected route need to include sign in token as authorization bearer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Protected route need to include sign in token as authorization bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/addresses/wards": {
            "get": {
                "tags": [
                    "AddressWard"
                ],
                "summary": "Get all address ward",
                "description": "This request is used to get all address ward when user use the App. This request is using MySql database.",
                "operationId": "1ac99953dcbd32e88cc5c880e76f77c9",
                "responses": {
                    "200": {
                        "description": "Address ward fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Get List Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AddressWard"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Protected route need to include sign in token as authorization bearer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Protected route need to include sign in token as authorization bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/posts/types": {
            "get": {
                "tags": [
                    "PostType"
                ],
                "summary": "Get all post type",
                "description": "This request is used to get all post type when user use the App. This request is using MySql database.",
                "operationId": "8112183090a7c26d48189489ffe510a3",
                "responses": {
                    "200": {
                        "description": "Post type fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Get List Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PostType"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Protected route need to include sign in token as authorization bearer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Protected route need to include sign in token as authorization bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AddressDistrict": {
                "required": [
                    "id",
                    "isDisable",
                    "createdAt",
                    "updatedAt",
                    "name",
                    "code",
                    "description",
                    "provinceCode"
                ],
                "properties": {
                    "id": {
                        "description": "Primary Key",
                        "type": "string",
                        "format": "uuid"
                    },
                    "isDisable": {
                        "description": "Indicates if the code is marked as a disable",
                        "type": "boolean"
                    },
                    "createdAt": {
                        "description": "Timestamp when the admin was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "description": "Timestamp when the admin was updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "name": {
                        "description": "Name of the district",
                        "type": "string"
                    },
                    "code": {
                        "description": "Code of the district",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the district",
                        "type": "string"
                    },
                    "provinceCode": {
                        "description": "Code of the province",
                        "type": "string"
                    },
                    "province": {
                        "$ref": "#/components/schemas/AddressProvince"
                    },
                    "wards": {
                        "description": "Wards of the address",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AddressWard"
                        }
                    }
                },
                "type": "object"
            },
            "AddressProvince": {
                "required": [
                    "id",
                    "isDisable",
                    "createdAt",
                    "updatedAt",
                    "name",
                    "code",
                    "description"
                ],
                "properties": {
                    "id": {
                        "description": "Primary Key",
                        "type": "string",
                        "format": "uuid"
                    },
                    "isDisable": {
                        "description": "Indicates if the code is marked as a disable",
                        "type": "boolean"
                    },
                    "createdAt": {
                        "description": "Timestamp when the admin was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "description": "Timestamp when the admin was updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "name": {
                        "description": "Name of the province",
                        "type": "string"
                    },
                    "code": {
                        "description": "Code of the province",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the province",
                        "type": "string"
                    },
                    "districts": {
                        "description": "Districts of the address",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AddressDistrict"
                        }
                    }
                },
                "type": "object"
            },
...
            "AddressWard": {
                "required": [
                    "id",
                    "isDisable",
                    "createdAt",
                    "updatedAt",
                    "name",
                    "code",
                    "description",
                    "districtCode"
                ],
                "properties": {
                    "id": {
                        "description": "Primary Key",
                        "type": "string",
                        "format": "uuid"
                    },
                    "isDisable": {
                        "description": "Indicates if the code is marked as a disable",
                        "type": "boolean"
                    },
                    "createdAt": {
                        "description": "Timestamp when the admin was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "description": "Timestamp when the admin was updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "name": {
                        "description": "Name of the ward",
                        "type": "string"
                    },
                    "code": {
                        "description": "Code of the ward",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the ward",
                        "type": "string"
                    },
                    "districtCode": {
                        "description": "Code of the district",
                        "type": "string"
                    },
                    "district": {
                        "$ref": "#/components/schemas/AddressDistrict"
                    }
                },
                "type": "object"
            },
...
            "PostLanguage": {
                "required": [
                    "id",
                    "isDisable",
                    "createdAt",
                    "updatedAt",
                    "name",
                    "slug",
                    "description",
                    "content",
                    "language"
                ],
                "properties": {
                    "id": {
                        "description": "Primary Key",
                        "type": "string",
                        "format": "uuid"
                    },
                    "isDisable": {
                        "description": "Indicates if the code is marked as a disable",
                        "type": "boolean"
                    },
                    "createdAt": {
                        "description": "Timestamp when the admin was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "description": "Timestamp when the admin was updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "name": {
                        "description": "Name of the post language",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Slug of the post language",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the post language",
                        "type": "string"
                    },
                    "content": {
                        "description": "Content HTML",
                        "type": "string"
                    },
                    "post": {
                        "$ref": "#/components/schemas/Post"
                    },
                    "language": {
                        "description": "Language of the post",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Post": {
                "required": [
                    "id",
                    "isDisable",
                    "createdAt",
                    "updatedAt",
                    "imageUrl",
                    "typeCode"
                ],
                "properties": {
                    "id": {
                        "description": "Primary Key",
                        "type": "string",
                        "format": "uuid"
                    },
                    "isDisable": {
                        "description": "Indicates if the code is marked as a disable",
                        "type": "boolean"
                    },
                    "createdAt": {
                        "description": "Timestamp when the admin was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "description": "Timestamp when the admin was updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "imageUrl": {
                        "description": "Image url of the post",
                        "type": "string"
                    },
                    "typeCode": {
                        "description": "Code of the type",
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/PostType"
                    },
                    "languages": {
                        "description": "Post language",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PostLanguage"
                        }
                    }
                },
                "type": "object"
            },
            "PostType": {
                "required": [
                    "id",
                    "isDisable",
                    "createdAt",
                    "updatedAt",
                    "name",
                    "code",
                    "description",
                    "postTypeId",
                    "postsCount"
                ],
                "properties": {
                    "id": {
                        "description": "Primary Key",
                        "type": "string",
                        "format": "uuid"
                    },
                    "isDisable": {
                        "description": "Indicates if the code is marked as a disable",
                        "type": "boolean"
                    },
                    "createdAt": {
                        "description": "Timestamp when the admin was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "description": "Timestamp when the admin was updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "name": {
                        "description": "Name of the type post",
                        "type": "string"
                    },
                    "code": {
                        "description": "Code of the type post",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the type post",
                        "type": "string"
                    },
                    "posts": {
                        "description": "List posts",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Post"
                        }
                    },
                    "postTypeId": {
                        "description": "Children post",
                        "type": "string",
                        "format": "uuid"
                    },
                    "children": {
                        "description": "Children post",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PostType"
                        }
                    },
                    "postsCount": {
                        "description": "Count post",
                        "type": "number"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}